Convert styles.scss to Sass syntax
As a first step in overhauling the styling of DadBot, the first step is to
convert the previous use of the .scss syntax to the more preferred Sass syntax.
By doing so, it will allow for easier tidying up in a future piece of work when
the styling will be improved.

This commit also involves a small change to `gulpfile.js` to reflect the two
different styles of stylesheet in the repo and to update the folder structure
accordingly.