Add `gulp-autoprefixer`
This commit adds the package `gulp-autoprefixer` to the project and uses it in
the `sass` gulp task when the CSS is compiled and generated.

By using `gulp-autoprefixer` it will automatically add the correct vendor
prefixes to certain CSS properties without having to write them in our Sass
files.

This improvement will allow for better cross-browser compatibility and reduce
clutter in the Sass files themselves.