DadBot 2.0
This commit contains the all-new source files for DadBot 2.0. DadBot has been
completely re-written from the ground up and is now built using React and
`create-react-app` which allows the maintaining and developing of DadBot to be
easier than ever.

DadBot 2.0 now also comes the following improvements in it's source code and
visuals:
- Support for more pleasing UI when loading a new random joke
- Update iPhone mock-up showing Facebook Messenger
- A better designed mobile experience
- A much lower CSS and JS bundle size due in part to code splitting and a leaner
stack

DadBot 2.0 also represents the phenomenal growth I have personally achieved as a
developer since releasing DadBot 1.0 in 2017. Hopefully, this growth will
continue to show in more improvements to DadBot in the very near future.