Handle Netlify redirect for `/random-joke`
This commit fixes an error in Netlify in which if a user arrived directly at the
`dadbot.com/random-joke` URL - they would be presented with a 404 error. This is
because there is not an index file located at /random-joke.

To prevent this from happening - we use a Netlify.toml file to configure our a
redirect back to the root URl and React Router handles the component that is
presented.